Title: 나선형 삼각형, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240818/나선형 삼각형/spira-triangle.cpp
@@ -7,7 +7,7 @@ int main() {
     int n;
     cin >> n;
 
-    int dp[101];
+    long long dp[101];
     dp[1] = 1;
     dp[2] = 1;
     dp[3] = 1;
